http://lists.maemo.org/pipermail/maemo-developers/2010-February/024534.html
The leds driver used on N900 - ledslp5523.c has a bug. The function
lp5523_set_brightness() which might be called in interrupt context, or
with interrupts disabled calls (eventually) i2c_smbus_xfer(), which
might sleep, which is not allowed. The function should use a work
queue, as in this patch.

The bug is not noticeable in usual working, since the leds brightness
is only set from userspace, so the function is only called in process
context. But trying to change one of the leds' trigger, results in
immediate "scheduling while atomic" oops.