NITdroid: kernel package debianised. NIT.04 released as a .deb
This package contains the kernel image, and the modules.
The modules are installed in Android, which must be mounted
on /and before installing this package. The kernel image
is installed in /boot/multiboot in Maemo.
The installer copies essential modules for booting up the kernel
into /lib/modules in Maemo, and also runs depmod.